requirements/base.txt
asgiref==3.8.1
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
cssselect2==0.7.0
decorator==4.4.2
defusedxml==0.7.1
Django==5.1.6
django-bootstrap5==24.3
django-ckeditor==6.7.2
django-js-asset==3.0.1
et_xmlfile==2.0.0
fire==0.7.0
flexcache==0.3
flexparser==0.4
fonttools==4.56.0
forex-python==1.8
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
lxml==5.3.1
markdown2==2.5.3
moviepy==1.0.3
numpy==2.2.3
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pandas==2.2.3
pdf2docx==0.5.8
pdf2image==1.17.0
pdfkit==1.0.0
pillow==11.1.0
Pint==0.24.4
platformdirs==4.3.6
proglog==0.1.10
pycparser==2.22
pydub==0.25.1
PyMuPDF==1.25.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.1
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
termcolor==2.5.0
tinycss2==1.4.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.2
webencodings==0.5.1
xlrd==2.0.1
xmltodict==0.14.2